change customer rank type to int

+++ Model/Customer.cs
@@ -16,7 +16,7 @@
         public decimal Score { get; set; }
 
 
-        public decimal Rank { get; set; }
+        public int Rank { get; set; }
 
 
         /// <summary>
